Build 0.0.2
I forgot to commit build 0.0.1, so sorry about that. Anyways, this build is pretty simple. Move with A and S, hold down D to have the ball bounce off of you, release D to do a charged shot, and press F to jump. The black circle tells you the radius of how far away you can do a charged shot.